verify_email

+++ src/controllers/user.controller.ts
@@ -2,8 +2,6 @@ import { Request, Response } from 'express'
 import Otp from '~/models/otp.model'
 import User from '~/models/user.model'
 import sendOTP from '~/utils/sendOTP'
-import brcypt from 'bcrypt'
-import jwt from 'jsonwebtoken'
 
 const register = async (req: Request, res: Response) => {
   try {
@@ -17,7 +15,7 @@ const register = async (req: Request, res: Response) => {
     }
 
     const code = Math.floor(100000 + Math.random() * 900000).toString()
-    const expiresAt = new Date(Date.now() + 5 * 60 * 1000)
+    const expiresAt = new Date(Date.now() + 1 * 60 * 1000)
 
     await Otp.deleteMany({ email })
     await Otp.create({ email, code, expiresAt })
@@ -35,21 +33,26 @@ const register = async (req: Request, res: Response) => {
 
 const verifyCode = async (req: Request, res: Response) => {
   try {
-    const { email, name, password, code } = req.body
-    const otpRecord = await Otp.findOne({ email, code })
-    if (!otpRecord) {
-      res.status(400).json({
-        message: 'Invalid verification code or expired'
-      })
-    }
-    const hashPassword = await brcypt.hash(password, 10)
-    const user = await User.create({ name, email, password: hashPassword })
-    const token = jwt.sign({ id: user._id }, process.env.JWT_SECRET as string)
+    const { email, code } = req.body
 
-    await User.create({ email, name, password })
+    const otpRecord = await Otp.findOne({ email, code })
+
+    if (!otpRecord) {
+      res.status(400).json({ message: 'Invalid verification code' })
+      return
+    }
+
+    if (!otpRecord.expiresAt || otpRecord.expiresAt < new Date()) {
+      res.status(400).json({ message: 'Verification code expired' })
+      return
+    }
+
     await Otp.deleteMany({ email })
 
-    res.json({ user, token, message: 'Registered successfully' })
+    res.status(200).json({
+      message: 'OTP verified successfully',
+      email
+    })
   } catch (error) {
     res.status(500).json({
       message: 'Internal server error',
